+++ src/ElmFormat/Render/Box.hs
@@ -2291,9 +2291,15 @@ formatType' elmVersion requireParens atype =
         AST.TypeVariable var ->
             line $ identifier $ formatVarName elmVersion var
 
-        AST.TypeConstruction ctor args ->
+        AST.TypeConstruction ctor args forceMultiline ->
+            let
+                join = 
+                    case forceMultiline of
+                        AST.ForceMultiline True -> AST.FASplitFirst
+                        AST.ForceMultiline False -> AST.FAJoinFirst AST.JoinAll
+            in
             ElmStructure.application
-                (AST.FAJoinFirst AST.JoinAll)
+                join
                 (formatTypeConstructor elmVersion ctor)
                 (map (formatHeadCommented $ formatType' elmVersion ForCtor) args)
                 |> (if args /= [] && requireParens == ForCtor then parens else id)
@@ -2301,8 +2307,8 @@ formatType' elmVersion requireParens atype =
         AST.TypeParens type' ->
           parens $ formatCommented (formatType elmVersion) type'
 
-        AST.TupleType types ->
-          ElmStructure.group True "(" "," ")" False (map (formatCommented (formatEolCommented $ formatType elmVersion)) types)
+        AST.TupleType types (AST.ForceMultiline forceMultiline) ->
+            ElmStructure.group True "(" "," ")" forceMultiline (map (formatCommented (formatEolCommented $ formatType elmVersion)) types)
 
         AST.RecordType base fields trailing multiline ->
             formatRecordLike