+++ tests/Parse/TypeTest.hs
@@ -13,7 +13,7 @@ import qualified Box
 import qualified Data.Text as Text
 
 
-pending = at 0 0 0 0 $ TupleType []
+pending = at 0 0 0 0 $ TupleType [] (ForceMultiline False) 
 
 
 example :: String -> String -> String -> TestTree
@@ -31,7 +31,7 @@ tests =
         , example "comments"
             "({-A-}a{-B-},{-C-}b{-D-})"
             "( {- A -} a {- B -}, {- C -} b {- D -} )\n"
-        , example "newlines" "(\n a\n ,\n b\n )" "( a, b )\n"
+        , example "newlines" "(\n a\n ,\n b\n )" "( a\n, b\n)\n"
         ]
 
     , testGroup "record type"