+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## main
+
+New features:
+  - newlines in tuple types are now preserved
+  - newlines in type constructor application are now preserved
+
+
 ## 0.8.3
 
 Bug fixes: